feat(data): refactor mnemonic models and remove init_db script
